Убираем директивы eslint, каждый класс исключения в отдельный файл
