+++ app.js
@@ -13,7 +13,7 @@ const { MONGODB_URL = 'mongodb://localhost:27017/newsdb' } = process.env;
 const index = require('./routes/index');
 
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-const { NotFoundError } = require('./modules/exceptions');
+const { NotFoundError } = require('./modules/exceptions/NotFoundError');
 
 mongoose.connect(MONGODB_URL, {
   useNewUrlParser: true,
@@ -59,7 +59,4 @@ app.use((err, req, res, next) => {
   next();
 });
 
-app.listen(PORT, () => {
-  /* eslint no-console: [ "error", {allow: ["log"] } ] */
-  console.log(`App listening on port ${PORT}`);
-});
+app.listen(PORT);
