--- modules/exceptions.js
@@ -1,43 +1,0 @@
-/* eslint max-classes-per-file: ["error", 5] */
-class BadRequestError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 400;
-  }
-}
-
-class UnauthorizedError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 401;
-  }
-}
-
-class BadPermitionsError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 403;
-  }
-}
-
-class NotFoundError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 404;
-  }
-}
-
-class AlreadyExistsError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 409;
-  }
-}
-
-module.exports = {
-  BadRequestError,
-  BadPermitionsError,
-  NotFoundError,
-  AlreadyExistsError,
-  UnauthorizedError,
-};
