+++ modules/exceptions/AlreadyExistsError.js
@@ -1,0 +1,8 @@
+class AlreadyExistsError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 409;
+  }
+}
+
+module.exports = AlreadyExistsError;
