+++ modules/exceptions/BadPermitionsError.js
@@ -1,0 +1,8 @@
+class BadPermitionsError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 403;
+  }
+}
+
+module.exports = BadPermitionsError;
