+++ modules/exceptions/UnauthorizedError.js
@@ -1,0 +1,8 @@
+class UnauthorizedError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 401;
+  }
+}
+
+module.exports = UnauthorizedError;
